Refactor paper layout JS components
