+++ src/scenes/papers/components/Header.js
@@ -2,7 +2,7 @@
 
 import { Component } from "react";
 import { Header, Menu, Segment } from "semantic-ui-react";
-import HeaderModal from "./HeaderModal";
+import HeaderModal from "./header/HeaderModal";
 import "./Header.css";
 
 export default class PapersHeader extends Component {
