+++ src/scenes/papers/components/Panel.css
