+++ src/scenes/papers/components/Sidebar.js
